message handling moved to Raiden
wip: moving the message handling to a central place before adding a
proper abstraction.
