+++ raiden/transfermanager.py
@@ -53,7 +53,6 @@ class TransferManager(object):
     def __init__(self, tokenmanager):
         self.tokenmanager = tokenmanager
 
-        self.exchanges = dict()  #: mapping for pending exchanges
         self.endtask_transfer_mapping = dict()
 
     def transfer_async(self, amount, target, identifier=None):
@@ -143,125 +142,3 @@ class TransferManager(object):
         task.start()
 
         return async_result
-
-    def on_mediatedtransfer_message(self, transfer):
-        if transfer.sender not in self.tokenmanager.partneraddress_channel:
-            if log.isEnabledFor(logging.WARN):
-                log.warn(
-                    'Received mediated transfer message from unknown channel.'
-                    'Sender: %s',
-                    pex(transfer.sender),
-                )
-            raise UnknownAddress
-
-        raiden = self.tokenmanager.raiden
-        token_address = self.tokenmanager.token_address
-
-        channel = self.tokenmanager.partneraddress_channel[transfer.sender]
-        if not channel.isopen:
-            if log.isEnabledFor(logging.WARN):
-                log.warn(
-                    'Received mediated transfer message from %s after channel closing',
-                    pex(transfer.sender),
-                )
-            raise TransferWhenClosed
-        channel.register_transfer(transfer)  # raises if the transfer is invalid
-
-        exchange_key = ExchangeKey(transfer.token, transfer.lock.amount)
-        if exchange_key in self.exchanges:
-            exchange = self.exchanges[exchange_key]
-
-            if log.isEnabledFor(logging.DEBUG):
-                log.debug(
-                    'EXCHANGE TRANSFER RECEIVED node:%s %s > %s hashlock:%s'
-                    ' from_token:%s from_amount:%s to_token:%s to_amount:%s [%s]',
-                    pex(self.tokenmanager.raiden.address),
-                    pex(transfer.sender),
-                    pex(self.tokenmanager.raiden.address),
-                    pex(transfer.lock.hashlock),
-                    pex(exchange.from_token),
-                    exchange.from_amount,
-                    pex(exchange.to_token),
-                    exchange.to_amount,
-                    repr(transfer),
-                )
-
-            exchange_task = ExchangeTask(
-                raiden,
-                from_mediated_transfer=transfer,
-                to_token=exchange.to_token,
-                to_amount=exchange.to_amount,
-                target=exchange.from_nodeaddress,
-            )
-            exchange_task.start()
-
-        elif transfer.target == self.tokenmanager.raiden.address:
-            if log.isEnabledFor(logging.DEBUG):
-                log.debug(
-                    'MEDIATED TRANSFER RECEIVED node:%s %s > %s hashlock:%s [%s]',
-                    pex(self.tokenmanager.raiden.address),
-                    pex(transfer.sender),
-                    pex(self.tokenmanager.raiden.address),
-                    pex(transfer.lock.hashlock),
-                    repr(transfer),
-                )
-
-            try:
-                self.tokenmanager.raiden.message_for_task(
-                    transfer,
-                    transfer.lock.hashlock
-                )
-            except UnknownAddress:
-                # assumes that the registered task(s) tooks care of the message
-                # (used for exchanges)
-                secret_request_task = EndMediatedTransferTask(
-                    raiden,
-                    token_address,
-                    transfer,
-                )
-                secret_request_task.start()
-
-        else:
-            if log.isEnabledFor(logging.DEBUG):
-                log.debug(
-                    'TRANSFER TO BE MEDIATED RECEIVED node:%s %s > %s hashlock:%s [%s]',
-                    pex(self.tokenmanager.raiden.address),
-                    pex(transfer.sender),
-                    pex(self.tokenmanager.raiden.address),
-                    pex(transfer.lock.hashlock),
-                    repr(transfer),
-                )
-
-            transfer_task = MediateTransferTask(
-                raiden,
-                token_address,
-                transfer,
-                0,  # TODO: calculate the fee
-            )
-            transfer_task.start()
-
-    def on_directtransfer_message(self, transfer):
-        if transfer.sender not in self.tokenmanager.partneraddress_channel:
-            if log.isEnabledFor(logging.WARN):
-                log.warn(
-                    'Received direct transfer message from unknown sender %s',
-                    pex(transfer.sender),
-                )
-            raise UnknownAddress
-
-        channel = self.tokenmanager.partneraddress_channel[transfer.sender]
-
-        if not channel.isopen:
-            if log.isEnabledFor(logging.WARN):
-                log.warn(
-                    'Received direct transfer message from %s after channel closing',
-                    pex(transfer.sender),
-                )
-            raise TransferWhenClosed
-        channel.register_transfer(transfer)
-
-    def on_exchangerequest_message(self, message):
-        # if matches any own order
-        # if signed for me and fee:
-            # broadcast
-        pass
